Fitness is a double instead of an int Rename geneFrom into withGeneFrom Implement Tour and Tours In simple GA, directly mutate the fresh individual instead of mutating the whole population in a second loop.

+++ src/main/java/com/tpeyrard/ga/FitnessComputation.java
@@ -1,7 +1,7 @@
 package com.tpeyrard.ga;
 
 public interface FitnessComputation {
-    int getFitness(Individual individual);
+    double getFitness(Individual individual);
 
     int maxFitness();
 }
